requirements.txt
aiohttp==3.6.3
aiohttp-cors==0.7.0
aiohttp-debugtoolbar==0.6.0
aiohttp-jinja2==1.2.0
aiohttp-security==0.4.0
aiohttp-session==2.9.0
aioredis==1.3.1
aiosmtplib~=1.1.6
aiosqlite==0.16.0
async-timeout==3.0.1
asyncpg==0.23.0
atlastk==0.12.2
atomicwrites==1.4.0
attrs==20.2.0
cchardet==2.1.6
certifi==2020.12.5
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.4.7
dnspython==2.1.0
email-validator==1.1.2
function==1.2.0
hiredis==1.1.0
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
iso8601==0.1.13
Jinja2==2.11.2
jsonschema==3.2.0
loguru==0.5.3
lxml==4.6.3
MarkupSafe==1.1.1
marshmallow==3.9.1
more-itertools==8.5.0
multidict==4.7.6
overloading==0.5.0
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.9.1
py==1.9.0
pycparser==2.20
pydantic==1.8.1
pyparsing==2.4.7
PyPika==0.44.0
pypika-tortoise==0.1.0
pyrsistent==0.17.3
pytest==6.1.1
pytoml==0.1.21
pytz==2020.4
PyJWT==2.1.0
requests==2.24.0
six==1.15.0
svgutils==0.3.4
toml==0.10.2
tortoise-orm==0.17.2
typing-extensions==3.7.4.3
urllib3==1.25.11
wcwidth==0.2.5
yarl==1.5.1
zipp==3.3.0
git+https://git.fly-d.ru/flydata/api_utils.git@0.2.1
